requirements.txt
Flask>=0.10.1
Flask-OAuthlib>=0.9.1
Flask-SQLAlchemy>=2.1
